+++ src/components/pages/Login.tsx
@@ -1,103 +1,5 @@
-import { loginFormSchema, LoginFormSchema } from "@/schema/loginFormSchema";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "../ui/form";
-import { Button } from "../ui/button";
-import { Input } from "../ui/input";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "../ui/card";
-import { useAuth } from "@/hooks/useAuth";
-import { LuLoader2 } from "react-icons/lu";
-import { ForgotPasswordForm } from "../organisms/ForgotPasswordForm";
+import { LoginForm } from "../organisms/LoginForm";
 
 export const Login = () => {
-  const { login, isLoading } = useAuth();
-
-  const form = useForm<LoginFormSchema>({
-    mode: "onChange",
-    resolver: zodResolver(loginFormSchema),
-    defaultValues: {
-      email: "",
-      password: "",
-    },
-  });
-
-  const handleSubmit = (user: LoginFormSchema) => {
-    login(user);
-    form.reset();
-  };
-
-  return (
-    <Card className="mx-auto max-w-sm">
-      <CardHeader>
-        <CardTitle className="text-2xl">ログイン</CardTitle>
-        <CardDescription>Eメールとパスワードを入力してください</CardDescription>
-      </CardHeader>
-      <CardContent>
-        <Form {...form}>
-          <form
-            onSubmit={form.handleSubmit(handleSubmit)}
-            className="space-y-8"
-          >
-            <FormField
-              control={form.control}
-              name="email"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Eメール</FormLabel>
-                  <FormControl>
-                    <Input placeholder="university@example.com" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="password"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>パスワード</FormLabel>
-                  <FormControl>
-                    <Input type="password" {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <Button
-              type="submit"
-              disabled={!form.formState.isValid || isLoading}
-              className="w-full"
-            >
-              {isLoading ? (
-                <>
-                  <LuLoader2 className="mr-2 h-4 w-4 animate-spin" />
-                  <span>Please Wait</span>
-                </>
-              ) : (
-                "ログイン"
-              )}
-            </Button>
-          </form>
-        </Form>
-      </CardContent>
-      <CardFooter>
-        <ForgotPasswordForm />
-      </CardFooter>
-    </Card>
-  );
+  return <LoginForm />;
 };
